frontend/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
python-jose>=3.3.0
python-dotenv>=1.0.0
